+++ src/calc.h
@@ -1,6 +1,6 @@
 #ifndef CALC_H
 #define CALC_H
 
-double suma(int a, int b);
+double suma(double a, double b);
 
 #endif // CALC_H
